requirements.txt
# Our favorite web framework
Django==1.9.5

# Read sensitive settings from the environment
python-decouple==3.0

# Ease path/files operations
Unipath==1.1

# 12fator-like database URL parser
dj-database-url==0.4.0

# Allow tests to be fun
pytest==2.9.1
pytest-django==2.9.1

# Allow to use .jade templates
pyjade==4.0.0
